Move package update handling to its own module
