+++ Editor/Scripts/PackageUpdate.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07efc3d62d8b5c444b4ba81213d2c9f2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
