+++ Editor/Scripts/PackageUpdate/PackageUpdateHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54e423ed4b5d11e42b2ed10c62b68030
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
